* added manual execution of the loader * updated README with reference to Akamai's official docs

+++ loader.js
@@ -1,4 +1,4 @@
-function getBehaviorSrc(name, path, description = "") {
+function _getBehaviorSrc(name, path, description = "") {
 	return fetch(path)
 	.then(response => response.text())
 	.then(xml => {return {
@@ -10,24 +10,29 @@ function getBehaviorSrc(name, path, description = "") {
 	}});
 }
 
-function createBehavior(name, path, asCustomOverride = false) {
-	return getBehaviorSrc(name, path)
+function _addBehavior(name, path, isCustomOverride = false, dryrun = false) {
+    console.log(`${dryrun? "// TODO: " : ""}addBehavior("${name}");`);
+    return _getBehaviorSrc(name, path)
         .then(behaviorSrc => {
-            let apiPrefix = asCustomOverride ? "custom-overrides" : "custom-behaviors";
+            if (dryrun) return true;
+
+            let apiPrefix = isCustomOverride ? "custom-overrides" : "custom-behaviors";
             return fetch(`https://control.akamai.com/papi/v0/${apiPrefix}`,
-            {
-                method: 'POST',
-                credentials: "same-origin",
-                body: JSON.stringify(behaviorSrc),
-                headers: new Headers({"Content-Type": "application/json",
-                    "PAPI-Use-Prefixes": "false"})
-            })
-    	});
+                {
+                    method: 'POST',
+                    credentials: "same-origin",
+                    body: JSON.stringify(behaviorSrc),
+                    headers: new Headers({"Content-Type": "application/json",
+                        "PAPI-Use-Prefixes": "false"})
+                })
+        });
 }
 
-function deleteBehavior(behavior, asCustomOverride = false) {
-    let apiPrefix = asCustomOverride ? "custom-overrides" : "custom-behaviors";
+function _deleteBehavior(behavior, isCustomOverride = false, dryrun = false) {
+    console.log(`${dryrun? "// TODO: " : ""}deleteBehavior("${behavior.name}");`);
+    if (dryrun) return true;
 
+    let apiPrefix = isCustomOverride ? "custom-overrides" : "custom-behaviors";
     return fetch(`https://control.akamai.com/papi/v0/${apiPrefix}/${behavior.behaviorId}`,
         {
             method: 'DELETE',
@@ -36,8 +41,8 @@ function deleteBehavior(behavior, asCustomOverride = false) {
         })
 }
 
-function excludeExistingBehaviors(behaviors, asCustomOverride = false) {
-    let apiPrefix = asCustomOverride ? "custom-overrides" : "custom-behaviors";
+function excludeExistingBehaviors(behaviors, isCustomOverride = false) {
+    let apiPrefix = isCustomOverride ? "custom-overrides" : "custom-behaviors";
 
     return fetch(`https://control.akamai.com/papi/v0/${apiPrefix}`, {credentials: "same-origin"})
         .then(response => response.json())
@@ -49,8 +54,8 @@ function excludeExistingBehaviors(behaviors, asCustomOverride = false) {
         })
 }
 
-function matchExistingBehaviors(behaviors, asCustomOverride = false) {
-    let apiPrefix = asCustomOverride ? "custom-overrides" : "custom-behaviors";
+function _matchExistingBehaviors(behaviors, isCustomOverride = false) {
+    let apiPrefix = isCustomOverride ? "custom-overrides" : "custom-behaviors";
 
     return fetch(`https://control.akamai.com/papi/v0/${apiPrefix}`, {credentials: "same-origin"})
         .then(response => response.json())
@@ -64,47 +69,60 @@ function matchExistingBehaviors(behaviors, asCustomOverride = false) {
         })
 }
 
-function createAllCustomBehaviors(asCustomOverride = false) {
-    let apiPrefix = asCustomOverride ? "/overrides" : "";
-
-    return fetch(`https://api.github.com/repos/colinbendell/pm-custom-behaviors/contents${apiPrefix}`)
-        .then(response => response.json())
-        .then(data => {
-            let behaviorList = new Map();
-            data.forEach(behaviorFile => {
-                if (/\.xml$/.test(behaviorFile.name) && behaviorFile.size > 0)
-                    behaviorList.set(behaviorFile.name.replace('.xml', ''), behaviorFile.download_url)
-            });
-            return behaviorList;
-        })
-        .then(behaviors => excludeExistingBehaviors(behaviors, asCustomOverride))
-        .then(behaviors => Promise.all([...behaviors].map(val => {createBehavior(val[0], val[1], asCustomOverride)})));
+function _addAllCustomBehaviors(dryrun = false, filter = null) {
+    return Promise.all([true, false].map(isCustomOverride => 
+        fetch(`https://api.github.com/repos/colinbendell/pm-custom-behaviors/contents${isCustomOverride ? "/overrides" : ""}`)
+            .then(response => response.status !== 200 ? Promise.reject(response) : response.json())
+            .then(data => {
+                let behaviorList = new Map();
+                data.forEach(behaviorFile => {
+                    if (/\.xml$/.test(behaviorFile.name)
+                        && (!filter || new RegExp(filter).test(behaviorFile.name))
+                        && behaviorFile.size > 0)
+                        behaviorList.set(behaviorFile.name.replace('.xml', ''), behaviorFile.download_url)
+                });
+                return behaviorList;
+            })
+            .then(behaviors => excludeExistingBehaviors(behaviors, isCustomOverride))
+            .then(behaviors => Promise.all([...behaviors].map(val => {_addBehavior(val[0], val[1], isCustomOverride, dryrun)})))
+        )
+    );
 }
 
-function deleteAllCustomBehaviors(asCustomOverride = false) {
-    let apiPrefix = asCustomOverride ? "/overrides" : "";
+function _deleteAllCustomBehaviors(dryrun = false, filter = null) {
+    return Promise.all([true, false].map(isCustomOverride =>
+        fetch(`https://api.github.com/repos/colinbendell/pm-custom-behaviors/contents${isCustomOverride ? "/overrides" : ""}/deprecated`)
+            .then(response => response.status !== 200 ? Promise.reject(response) : response.json())
+            .then(data => {
+                let behaviorSet = new Set();
 
-    return fetch(`https://api.github.com/repos/colinbendell/pm-custom-behaviors/contents${apiPrefix}/deprecated`)
-        .then(response => response.json())
-        .then(data => {
-            let behaviorSet = new Set();
-
-            data.forEach(behaviorFile => {
-                if (/\.xml$/.test(behaviorFile.name))
-                    behaviorSet.add(behaviorFile.name.replace('.xml', ''))
-            });
-            return behaviorSet;
-        })
-        .then(behaviors => matchExistingBehaviors(behaviors, asCustomOverride))
-        .then(behaviors => Promise.all(behaviors.map(o => {deleteBehavior(o, asCustomOverride)})));
+                data.forEach(behaviorFile => {
+                    if (/\.xml$/.test(behaviorFile.name)
+                        && (!filter || new RegExp(filter).test(behaviorFile.name)))
+                        behaviorSet.add(behaviorFile.name.replace('.xml', ''))
+                });
+                return behaviorSet;
+            })
+            .then(behaviors => _matchExistingBehaviors(behaviors, isCustomOverride))
+            .then(behaviors => Promise.all(behaviors.map(o => {_deleteBehavior(o, isCustomOverride, dryrun)})))
+        )
+    );
 }
 
-function main() {
-    createAllCustomBehaviors();
-    deleteAllCustomBehaviors();
-    createAllCustomBehaviors(true);
-    deleteAllCustomBehaviors(true);
+function sync(dryrun = false) {
+    return _addAllCustomBehaviors(dryrun)
+        .then (() => _deleteAllCustomBehaviors(dryrun))
+        .then(() => dryrun ? console.log('//TODO: run `sync();` to automatically perform the above') : '')
+        .catch(e => console.error(e))
 }
 
-main();
+function addBehavior(name) {
+    return _addAllCustomBehaviors(false, name);
+}
+
+function deleteBehavior(name) {
+    return _deleteAllCustomBehaviors(false, name);
+}
+
+sync(true);
